More corrections
There still were some errors on a fresh install. Trying to simplify the
cloning an build of some tools. No need to remove old clones, clone, and
then run checkout as command. The git module can handle checking out a
specific version.

The tasks files for the gpg role was added in the wrong place.

Tried to fix the installation of vim plugins as part of running the
playbook. May be necessary to run :CocUpdate/CocUpdateSync as well.